CRITICAL ISSUE Jobs queued behind a stuck job in the Invoking queue for 7-10 days, causing a complete system failure without throwing an exception that would stop the server.
RESOLUTION:
Server rebooted by ITG on 7/30, see JDA Helpdesk ticket #900893.
This restored remote access to the server.

NEXT STEPS:
A Fagan process is needed to remove hung jobs, or a failover needs to be implemented to escalate an exception
to the server level after a certain number of jobs have failed.